requirements.txt
beautifulsoup4==4.9.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
google-api-core==1.22.1
google-api-python-client==1.11.0
google-auth==1.21.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
GoogleNews==1.4.1
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.24.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
wikipedia==1.4.0
Wikipedia-API==0.5.4
